refactor: 抽离公共的测试world，减少copy-paste问题

+++ src/common/world/storage/db/RocksDBConfig.hpp
@@ -165,7 +165,7 @@ struct RocksDBConfig {
         // 统计
         if (enableStatistics) {
             options.statistics = rocksdb::CreateDBStatistics();
-            options.stats_dump_period_sec = statisticsDumpPeriodMs / 1000;
+            options.stats_dump_period_sec = static_cast<unsigned int>(statisticsDumpPeriodMs / 1000);
         }
 
         return options;
